packages/apps/apps/cerebro/requirements-min.txt
# Minimal requirements for backend to run
# WIP: use this as a reference to build a minimal docker image

fastapi==0.128.0
uvicorn[standard]==0.39.0
pydantic==2.12.5
python-multipart==0.0.21
itsdangerous==2.2.0

python-socketio==5.16.0
python-jose==3.5.0
cryptography
bcrypt==5.0.0
argon2-cffi==25.1.0
PyJWT[crypto]==2.10.1
authlib==1.6.6

requests==2.32.5
aiohttp==3.13.2
async-timeout
aiocache
aiofiles
starlette-compress==1.6.1
httpx[socks,http2,zstd,cli,brotli]==0.28.1
starsessions[redis]==2.2.1

sqlalchemy==2.0.45
alembic==1.17.2
peewee==3.18.3
peewee-migrate==1.14.3

pycrdt==0.12.44
redis

APScheduler==3.11.2
RestrictedPython==8.1

loguru==0.7.3
asgiref==3.11.0

mcp==1.25.0
openai

langchain==1.2.0
langchain-community==0.4.1
langchain-classic==1.0.1
langchain-text-splitters==1.1.0

fake-useragent==2.2.0

chromadb==1.4.0
black==25.12.0
pydub
chardet==5.2.0
